Completed-ish input/output
Got the input and output working with the raspberry pi.
One file handles the input and the other handles the output.
Really I should have a seperate file pulled in that associates what variables i need and what pins i need and maybe dataref info like maybe the on off is 0 and 2 not 0 and 1.  Something variable i donno ye.